mcp-server/requirements.txt
fastapi
uvicorn
langchain
python-dotenv
pydantic
langchain-google-genai
